Adds function to estimate servo panning time
