+++ oliphaunt/servolib.cpp
@@ -1,5 +1,7 @@
 #include "servolib.h"
+
 #include <Arduino.h>
+#include <math.h>
 #include "utility.h"
 
 /** Drive Servos **/
@@ -27,10 +29,10 @@ void driveServoRightCmd(byte command) {
 void driveServoLeft(float setting) {
     byte cmd = SERVO_DRIVE_LEFT_PWM_STOP;
     if(setting > 0) {
-        cmd = (byte) LINSCALE(setting, 0.0, +1.0, SERVO_DRIVE_LEFT_PWM_STOP, SERVO_DRIVE_LEFT_PWM_MAX);
+        cmd = (byte) round(LINSCALE(setting, 0.0, +1.0, SERVO_DRIVE_LEFT_PWM_STOP, SERVO_DRIVE_LEFT_PWM_MAX));
     }
     else if (setting < 0) {
-        cmd = (byte) LINSCALE(setting, -1.0, 0.0, SERVO_DRIVE_LEFT_PWM_MIN, SERVO_DRIVE_LEFT_PWM_STOP);
+        cmd = (byte) round(LINSCALE(setting, -1.0, 0.0, SERVO_DRIVE_LEFT_PWM_MIN, SERVO_DRIVE_LEFT_PWM_STOP));
     }
     driveServoLeftCmd(cmd);
 }
@@ -38,10 +40,10 @@ void driveServoLeft(float setting) {
 void driveServoRight(float setting) {
     byte cmd = SERVO_DRIVE_RIGHT_PWM_STOP;
     if(setting > 0) {
-        cmd = (byte) LINSCALE(setting, 0.0, +1.0, SERVO_DRIVE_RIGHT_PWM_STOP, SERVO_DRIVE_RIGHT_PWM_MAX);
+        cmd = (byte) round(LINSCALE(setting, 0.0, +1.0, SERVO_DRIVE_RIGHT_PWM_STOP, SERVO_DRIVE_RIGHT_PWM_MAX));
     }
     else if (setting < 0) {
-        cmd = (byte) LINSCALE(setting, -1.0, 0.0, SERVO_DRIVE_RIGHT_PWM_MIN, SERVO_DRIVE_RIGHT_PWM_STOP);
+        cmd = (byte) round(LINSCALE(setting, -1.0, 0.0, SERVO_DRIVE_RIGHT_PWM_MIN, SERVO_DRIVE_RIGHT_PWM_STOP));
     }
     driveServoRightCmd(cmd);    
 }
@@ -50,7 +52,7 @@ void driveServoRight(float setting) {
 static float panning_setpoint;
 
 void setPanningServo(float angle) {
-    byte cmd = (byte) LINSCALE(angle, -90, +90, SERVO_PANNING_PWM_MIN, SERVO_PANNING_PWM_MAX);
+    byte cmd = (byte) round(LINSCALE(angle, -90, +90, SERVO_PANNING_PWM_MIN, SERVO_PANNING_PWM_MAX));
     analogWrite(SERVO_PANNING_PWM_PIN, cmd);
     panning_setpoint = angle;
 }
@@ -59,3 +61,7 @@ float getPanningServoSetpoint() {
     return panning_setpoint;
 }
 
+unsigned long estimatePanningTime(float target_angle) {
+    float arc = getShortestArc(panning_setpoint, target_angle);
+    return (unsigned long) ceil(arc*SERVO_PANNING_SPEED);
+}
